1) Fixed XML-mapping of model a bit. 2) Added classes for savind data in DB. 3) Implemented data migration from parser to db-model.

+++ postgreXML/src/main/java/com/komarov/postgresXML/model/Face.java
@@ -4,43 +4,43 @@ import javax.xml.bind.annotation.XmlAttribute;
 import java.io.Serializable;
 
 public class Face implements Serializable {
-    private float v1;
-    private float v2;
-    private float v3;
+    private int v1;
+    private int v2;
+    private int v3;
 
     public Face() {
     }
 
-    public Face(float v1, float v2, float v3) {
+    public Face(int v1, int v2, int v3) {
         this.v1 = v1;
         this.v2 = v2;
         this.v3 = v3;
     }
 
     @XmlAttribute
-    public float getV1() {
+    public int getV1() {
         return v1;
     }
 
-    public void setV1(float v1) {
+    public void setV1(int v1) {
         this.v1 = v1;
     }
 
     @XmlAttribute
-    public float getV2() {
+    public int getV2() {
         return v2;
     }
 
-    public void setV2(float v2) {
+    public void setV2(int v2) {
         this.v2 = v2;
     }
 
     @XmlAttribute
-    public float getV3() {
+    public int getV3() {
         return v3;
     }
 
-    public void setV3(float v3) {
+    public void setV3(int v3) {
         this.v3 = v3;
     }
 }
